requirements.txt
PyYAML==3.10
distribute==0.6.28
httplib2==0.7.6
oauth2==1.5.211
python-twitter==0.8.2
simplejson==2.6.2
tweepy==1.12
wsgiref==0.1.2
